+++ lib/src/ref.dart
@@ -1,66 +1,87 @@
+import 'dart:async';
+
+import 'package:flutter/material.dart';
 import 'package:riverpie/src/listener.dart';
 import 'package:riverpie/src/notifier.dart';
 import 'package:riverpie/src/provider/provider.dart';
 
-/// "read" function type
-typedef RefRead = T Function<T>(BaseProvider<T> provider);
-
-/// "notifier" function type
-typedef RefNotifier = N Function<N extends BaseNotifier<T>, T>(
-    NotifierProvider<N, T> provider);
-
-/// "listen" function type
-typedef RefListen = T Function<T>(
-    BaseProvider<T> provider, ListenerCallback<T> listener);
-
 /// The base ref to read and notify providers.
 /// These methods can be called anywhere.
+/// Even within dispose methods.
+/// The primary difficulty is to get the [Ref] in the first place.
 abstract class Ref {
   /// Get the current value of a provider without listening to changes.
   T read<T>(BaseProvider<T> provider);
 
   /// Get the notifier of a provider.
   N notifier<N extends BaseNotifier<T>, T>(NotifierProvider<N, T> provider);
+
+  /// Listen for changes to a provider.
+  ///
+  /// Do not call this method during build as you
+  /// will create a new listener every time.
+  ///
+  /// You need to dispose the subscription manually.
+  Stream<NotifierEvent<T>> stream<N extends BaseNotifier<T>, T>(
+    NotifierProvider<N, T> provider,
+  );
 }
 
-/// The ref to watch providers (in addition to read and notify).
-/// Only call [watch] during build.
+/// The ref to watch providers (in addition to read, notify and stream).
 class WatchableRef extends Ref {
+  final Ref _root;
+  final State _state;
+
   @override
   T read<T>(BaseProvider<T> provider) {
-    return _read<T>(provider);
+    return _root.read<T>(provider);
   }
 
   @override
   N notifier<N extends BaseNotifier<T>, T>(NotifierProvider<N, T> provider) {
-    return _notifier<N, T>(provider);
+    return _root.notifier<N, T>(provider);
+  }
+
+  @override
+  Stream<NotifierEvent<T>> stream<N extends BaseNotifier<T>, T>(
+    NotifierProvider<N, T> provider,
+  ) {
+    return _root.stream<N, T>(provider);
   }
 
   /// Get the current value of a provider and listen to changes.
   /// The listener will be disposed automatically when the widget is disposed.
-  T watch<T>(BaseProvider<T> provider) {
-    return _watch<T>(provider);
-  }
+  /// Only call [watch] during build.
+  T watch<T>(
+    BaseProvider<T> provider, {
+    ListenerCallback<T>? listener,
+    bool Function(T prev, T next)? rebuildWhen,
+  }) {
+    if (provider is Provider<T>) {
+      // A normal provider is immutable, so we can return the value directly.
+      return _root.read(provider);
+    } else if (provider is NotifierProvider<BaseNotifier<T>, T>) {
+      // A notifier provider is mutable, so we also need to add a listener.
+      final notifier = _root.notifier(provider);
+      notifier.addListener(
+        _state,
+        ListenerConfig(
+          callback: listener,
+          selector: rebuildWhen,
+        ),
+      );
 
-  /// Get the current value of a provider and listen to changes.
-  /// In addition, a callback can be provided to be notified of changes.
-  /// The listener will be disposed automatically when the widget is disposed.
-  T listen<T>(BaseProvider<T> provider, ListenerCallback<T> listener) {
-    return _listen<T>(provider, listener);
-  }
+      // ignore: invalid_use_of_protected_member
+      return notifier.state;
+    }
 
-  final RefRead _read;
-  final RefNotifier _notifier;
-  final RefRead _watch;
-  final RefListen _listen;
+    throw Exception(
+        'Only [Provider] and [NotifierProvider] are supported, got: ${provider.runtimeType}');
+  }
 
   WatchableRef({
-    required RefRead read,
-    required RefNotifier notifier,
-    required RefRead watch,
-    required RefListen listen,
-  })  : _read = read,
-        _notifier = notifier,
-        _watch = watch,
-        _listen = listen;
+    required Ref root,
+    required State state,
+  })  : _root = root,
+        _state = state;
 }